apps/backend/requirements.txt
flake8>=7.0.0
pep8-naming>=0.11.1

Django>=5.0.2
djangorestframework>=3.14.0
django-cors-headers
python-decouple
python-dotenv
mysqlclient
    python-dotenv